+++ src/variants/model/likelihood/allelefreq_estimation.rs
@@ -1,0 +1,1 @@
+pub(crate) struct AlleleFreqEstimation
